improve migrations

+++ db/migrations/20141221205038_add_account_to_user.js
@@ -1,7 +1,7 @@
 'use strict';
 
 exports.up = function(knex, Promise) {
-    knex.schema.hasTable('users').then(function(exists) {
+    return knex.schema.hasTable('users').then(function(exists) {
         return knex.schema.table('users', function (table) {
             table.integer('account').defaultTo(0);
         })
@@ -9,7 +9,11 @@ exports.up = function(knex, Promise) {
 };
 
 exports.down = function(knex, Promise) {
-    return knex.schema.table('users', function (table) {
-        table.dropColumn('account');
+    return knex.schema.hasColumn('users','account').then(function(exists){
+        if(exists){
+            return knex.schema.table('users', function (table) {
+                table.dropColumn('account');
+            });
+        }
     });
 };
